pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dnd-music-tool"
version = "0.1.0"
description = "Backend tool for retrieving RPG music tracks by scene and genre"
authors = [{name = "DnD Music Tool"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
]

[tool.setuptools]
packages = ["app"]
include-package-data = true

[tool.setuptools.package-data]
app = []

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
